forgot to explain more about the previous commit
added a dataset variability study, which i will be running overnight for 100 iterations. meant to prove that a constellation's performance will not change significantly if it is simulated again on the same number of debris but a different dataset.
also added a constellation benchmark file, a cleaned up version of simulation.py meant to rapidly test constellations. rn working on simulating most requested confifgurations.
added comments in simulation.py, cleared the code, and fixed a horrible bug that overwrote deb and sat positions in subsequent batches, messing up the plots.
cleared a type in gaussian.py, deleted pso (old) and renamed pso copy to pso.
thats all folks